requirements.txt
flask==2.0.1
python-docx==0.8.11
boto3==1.26.137
python-dotenv==1.0.0
gunicorn==20.1.0
Werkzeug==2.0.1 